requirements.txt
flask==3.1.1
flask-cors==4.0.0
python-dotenv==1.1.0
duckduckgo-search==8.0.3
requests==2.32.4
gunicorn==21.2.0
Pillow==10.4.0
werkzeug==2.3.8
